Feature gate bom_table
