+++ crates/pcb-sch/Cargo.toml
@@ -26,8 +26,12 @@ starlark = { workspace = true }
 starlark_map = { workspace = true }
 allocative = { workspace = true }
 anyhow = { workspace = true }
-comfy-table = { workspace = true }
+comfy-table = { workspace = true, optional = true }
 starlark_syntax = { workspace = true }
+
+[features]
+default = []
+table = ["comfy-table"]
 
 [dev-dependencies]
 tempfile = { workspace = true }
