+++ crates/pcb/Cargo.toml
@@ -25,9 +25,9 @@ path = "src/main.rs"
 [dependencies]
 anyhow = { workspace = true }
 clap = { workspace = true }
-pcb-zen-core = { workspace = true }
+pcb-zen-core = { workspace = true, features = ["table"] }
 pcb-zen = { workspace = true }
-pcb-sch = { workspace = true }
+pcb-sch = { workspace = true, features = ["table"] }
 pcb-layout = { workspace = true }
 pcb-sim = { workspace = true }
 pcb-diode-api = { workspace = true, optional = true }
